refactor: add shared native components
