+++ apps/web/src/App.tsx
@@ -1,0 +1,5 @@
+import { HomeScreen } from '@dealer/shared';
+
+export default function App() {
+  return <HomeScreen />;
+}
