--- apps/web/src/pages/AppPage.tsx
@@ -1,10 +1,0 @@
-import Home from './HomePage';
-import { UiProvider } from '../context/UiContext';
-
-export default function App() {
-  return (
-    <UiProvider>                 
-      <Home />
-    </UiProvider>
-  );
-}
