--- apps/web/src/types/index.ts
@@ -1,7 +1,0 @@
-export interface CardDto {
-  id: number;
-  heading: string;
-  body: string[];
-  img: string;
-  cta: string;
-}
