+++ apps/web/tests/setup.ts
@@ -1,12 +1,7 @@
 import '@testing-library/jest-dom';
-import { vi, beforeAll, afterEach, afterAll } from 'vitest';
-import { server } from './mocks/server';
+import { vi } from 'vitest';
 
-// Vitest's node runtime lacks `window.matchMedia`, which some components
-// rely on. Define a minimal stub so tests can execute without throwing.
-
-// Ensure a global `window` object exists when running in a pure Node
-// environment.
+// Stub window.matchMedia for environments without it
 // @ts-expect-error: define `window` for Node runtime
 if (typeof window === 'undefined') global.window = {} as Window;
 
@@ -25,7 +20,3 @@ if (!window.matchMedia) {
     })),
   });
 }
-
-beforeAll(() => server.listen());
-afterEach(() => server.resetHandlers());
-afterAll(() => server.close());
