+++ apps/web/tests/smoke.test.ts
@@ -1,0 +1,7 @@
+import { describe, it, expect } from 'vitest';
+
+describe('smoke test', () => {
+  it('runs', () => {
+    expect(1).toBe(1);
+  });
+});
