+++ apps/web/tsconfig.app.json
@@ -1,7 +1,7 @@
 {
   "extends": ["../../tsconfig.base.json"],
   "compilerOptions": {
-    "types": ["vite/client"],
+    "types": ["vite/client", "react-native"],
     "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.app.tsbuildinfo",
     "target": "ES2022",
     "useDefineForClassFields": true,
