+++ apps/web/tsconfig.json
@@ -8,7 +8,8 @@
     "forceConsistentCasingInFileNames": true,
     "types": [
       "vitest/globals",
-      "node"
+      "node",
+      "react-native"
     ]
   },
   "files": [],
