+++ apps/web/tsconfig.test.json
@@ -1,7 +1,7 @@
 {
   "extends": "./tsconfig.app.json",
   "compilerOptions": {
-    "types": ["vitest/globals", "node"]
+    "types": ["vitest/globals", "node", "react-native"]
   },
   "include": ["tests", "src"]
 }
