+++ apps/web/vite.config.ts
@@ -1,20 +1,21 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
-import path from 'path'
-import { fileURLToPath, URL } from "node:url";
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
+import path from 'path';
+import { fileURLToPath, URL } from 'node:url';
 
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [react()],
   resolve: {
     alias: {
-      "@dealer/shared": fileURLToPath(new URL("../../packages/shared/src", import.meta.url)),
+      '@dealer/shared': fileURLToPath(new URL('../../packages/shared/src', import.meta.url)),
       '@': path.resolve(__dirname, './src'),
+      'react-native$': 'react-native-web',
     },
   },
   test: {
     globals: true,
     environment: 'jsdom',
-    setupFiles: './tests/setup.ts'
-  }
-})
+    setupFiles: './tests/setup.ts',
+  },
+});
