--- packages/shared/src/demo.ts
@@ -1,7 +1,0 @@
-export function formatPrice(cents: number, currency = "USD") {
-    return new Intl.NumberFormat("en-US", {
-      style: "currency",
-      currency
-    }).format(cents / 100);
-  }
-
