+++ packages/shared/src/index.ts
@@ -1,2 +1,4 @@
-export * from "./demo";
-export const SHARED_VERSION = "0.1.0";
+export { default as Button } from './components/Button';
+export { default as Card } from './components/Card';
+export { default as HomeScreen } from './screens/HomeScreen';
+export const SHARED_VERSION = '0.1.0';
