+++ packages/shared/src/screens/HomeScreen.tsx
@@ -1,15 +1,14 @@
 import { useState } from 'react';
 import { SafeAreaView, FlatList, StyleSheet } from 'react-native';
 import Card from '../components/Card';
-import type { CardData } from '../data/cards';
-import { cards } from '../data/cards';
+import { cards, CardContent } from '../data/cards';
 
 export default function HomeScreen() {
   const [selectedId, setSelectedId] = useState<number | null>(null);
 
   return (
     <SafeAreaView style={styles.container}>
-      <FlatList<CardData>
+      <FlatList<CardContent>
         data={cards}
         keyExtractor={(item) => item.id.toString()}
         renderItem={({ item }) => (
